listview.builder -> separated + scroll physics
