+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:provider/provider.dart';
-import 'package:snews/hidden_drawer.dart';
 import 'package:snews/src/feature/presentation/bloc/all_news_bloc.dart';
+import 'package:snews/src/feature/presentation/widgets/hidden_drawer.dart';
 
 void main() {
   runApp(const MainApp());
