+++ lib/src/feature/presentation/screens/main_screen.dart
@@ -27,77 +27,83 @@ class _MainScreenState extends State<MainScreen> {
           return MaterialApp(
             debugShowCheckedModeBanner: false,
             home: Scaffold(
-              body: ListView.builder(
+              body: ListView.separated(
+                  separatorBuilder: (BuildContext context, int index) =>
+                      const Divider(),
+                  physics: const BouncingScrollPhysics(decelerationRate: ScrollDecelerationRate.normal),
                   itemCount: state.loaded.results!.length,
                   itemBuilder: (context, index) {
                     return Padding(
-                      padding: const EdgeInsets.fromLTRB(10, 16, 10, 0),
+                      padding: const EdgeInsets.fromLTRB(10, 0, 10, 0),
                       child: Container(
                         decoration: const BoxDecoration(
                           color: Colors.white,
-                          border: Border(
-                            bottom: BorderSide(style: BorderStyle.solid),
-                          ),
-                          // borderRadius: BorderRadius.all(
-                          //   Radius.circular(10),
+                          // border: Border(
+                          //   bottom: BorderSide(style: BorderStyle.solid),
                           // ),
                         ),
-                        child: Wrap(
-                          direction: Axis.horizontal,
-                          children: [
-                            SizedBox(
-                              height: 150,
-                              width: MediaQuery.of(context).size.width / 2.2,
-                              child: Padding(
-                                padding: const EdgeInsets.fromLTRB(0, 0, 0, 15),
-                                child: Image.network(state
-                                    .loaded.results![index].multimedia![0].url
-                                    .toString()),
+                        child: Padding(
+                          padding: const EdgeInsets.fromLTRB(0, 15, 0, 0),
+                          child: Wrap(
+                            direction: Axis.horizontal,
+                            children: [
+                              SizedBox(
+                                height: 150,
+                                width: MediaQuery.of(context).size.width / 2.2,
+                                child: Padding(
+                                  padding:
+                                      const EdgeInsets.fromLTRB(0, 0, 0, 15),
+                                  child: Image.network(
+                                    state.loaded.results![index].multimedia![0]
+                                        .url
+                                        .toString(),
+                                    fit: BoxFit.fill,
+                                  ),
+                                ),
                               ),
-                            ),
-                            SizedBox(
-                              width: 194,
-                              child: Column(
-                                children: [
-                                  Padding(
-                                    padding:
-                                        const EdgeInsets.fromLTRB(10, 5, 10, 5),
-                                    child: Text(
-                                      state.loaded.results![index].title
-                                          .toString(),
-                                      overflow: TextOverflow.clip,
-                                      style: const TextStyle(
-                                        fontWeight: FontWeight.bold,
-                                        color: Color.fromRGBO(18, 18, 18, 1),
+                              SizedBox(
+                                width: 194,
+                                child: Column(
+                                  children: [
+                                    Padding(
+                                      padding: const EdgeInsets.fromLTRB(
+                                          10, 0, 10, 5),
+                                      child: Text(
+                                        state.loaded.results![index].title
+                                            .toString(),
+                                        overflow: TextOverflow.clip,
+                                        style: const TextStyle(
+                                          fontWeight: FontWeight.bold,
+                                          color: Color.fromRGBO(18, 18, 18, 1),
+                                        ),
                                       ),
                                     ),
-                                  ),
-                                  Padding(
-                                    padding: const EdgeInsets.fromLTRB(
-                                        10, 5, 10, 10),
-                                    child: Text(
-                                      state.loaded.results![index].abstract
-                                          .toString(),
-                                      overflow: TextOverflow.clip,
-                                      style: const TextStyle(
-                                        color: Color.fromRGBO(90, 90, 90, 1),
+                                    Padding(
+                                      padding: const EdgeInsets.fromLTRB(
+                                          10, 5, 10, 10),
+                                      child: Text(
+                                        state.loaded.results![index].abstract
+                                            .toString(),
+                                        overflow: TextOverflow.ellipsis,
+                                        style: const TextStyle(
+                                          color: Color.fromRGBO(90, 90, 90, 1),
+                                        ),
+                                        maxLines: 4,
                                       ),
                                     ),
-                                  ),
 
-                                  // Хотел под текстом с описанием новости указать дату публикации,
-                                  //но почему-то выдает null
+                                    // TODO: Хотел под текстом с описанием новости указать дату публикации,
+                                    //но почему-то выдает null
 
-                                  // Text(
-                                  //   state.loaded.results![index].publishedDate
-                                  //       .toString(),
-                                  // ),
-
-
-                                ],
+                                    // Text(
+                                    //   state.loaded.results![index].publishedDate
+                                    //       .toString(),
+                                    // ),
+                                  ],
+                                ),
                               ),
-                            ),
-                          ],
+                            ],
+                          ),
                         ),
                       ),
                     );
